Fixes broken questionnaire preview  when running using eq-compose.
The publisher application was not binding to address 0.0.0.0 when running
using docker compose, since the HOSTNAME environment variable was being
set by the docker engine and resolving to the container name.

This resulted in connection refused errors being reported by publisher
when the survey launcher attempted to request the survey JSON.

This fix, changes publisher to always bind to 0.0.0.0.