Add ability to route based on number into publisher.
This change refactors some of the existing logic for building up routing conditions and introduces some new logic for building routing conditions for numeric answers.

Whilst working on the tests I've been refactoring to try and communicate the business logic a bit more clearly. I've also written some tests to ensure that the existing behaviour for routing based on a checkbox is not broken by the new changes.

The logic for calculating the options for both the current condition and all conditions has been pushed down into the class. This reduces complexity and the number of arguments required to be passed down by the caller.